u-boot: Load kernel from current rootfs on TX2
Switch to loading the kernel from the current
active root filesystem on the TX2 to fix altboot
rollback issues where modules don't match the rolledback
version. Take into consideration sd card flasher images.

If rollback altboot is performed to an incompatible previous
version, there's not much that can be done unfortunately,
because the old kernel is incompatible with the new
partition layout and contents, and uboot cannot recreate
the previous partition table.

Signed-off-by: Alexandru Costache <alexandru@balena.io>